app/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
RPi.GPIO>=0.7.1
Pillow>=10.1.0
google-auth>=2.25.2
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
rpi_ws281x>=5.0.0
pygame>=2.5.2
flask>=3.0.0
flask-cors>=4.0.0
PyJWT>=2.8.0
SQLAlchemy>=2.0.23
PyMySQL>=1.1.0
cryptography>=41.0.7
Werkzeug>=3.0.1
email-validator>=2.1.0.post1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
PyQt5>=5.15.9
qrcode>=7.4.2 